Correct behaviour and alignment for buttons inside modifier list show_in_editmode and show_on_cage buttons behave similary to the regular UI now. At least almost:
Some modifiers show show_on_cage in the regular UI only if, for example,
an object to use for deforming is specified. Eg. Armatature modifier
requires an armature object to be specified in order to show the button.
This in not taken into account. In those cases, the button is always
visible. It's easier to achieve and hardly makes a difference.

Same behaviour for those buttons in modifier settings will be added
next.