Implement optional stack layout
Includes an operator for toggling all modifier panels. Gizmo settings
are in the Modifier Extras menu.

Also, ml_modifier_active_index is now overridable and uses
modifier.is_active. This allowed to remove the
window_manager.ml_active_object_active_modifier propery which was used
to workaround the issue with the modifier list not being able to set the
object.ml_modifier_active index for an object with library override.

This is currently only working in Properties Editor because of a bug in
Blender: https://developer.blender.org/T88655.