Fix error with lattice edit mode toggling when there is pinned object
Fixes an error when toggling lattice edit mode in Property Editor when
sidebar or popup is using a pinned object. It was caused by switching
from Property Editor to 3D View to get the active object (to restore the
active object after editing the lattice) in the viewport (rather than
the active object of Property Editor which can be another object that
has been pinned to Property Editor) and staying in that editor during
other operations also. The issue with that was that
ml_modifier_active_index uses a getter which uses get_ml_active_object
which in this case took the object from 3D View context instead of
Property Editor context. This caused the Edit Lattice operator to use a
modifier from a wrong object.

Fortunately, there is actually no need to swich the editor type at all
because context.active_object can be used to get the object that is
active in the viewport. Switching the editor type back to Property
Editor immediately after grabbing the active object should have also
worked.