Fix errors in console sometimes when reading ml_modifier_active_index
Before, the getter and the setter used get_ml_active_object instead of
accessing the object (self) directly and in some cases, for some reason,
when reading the active index, the context did not have the object
property, so get_ml_active_object caused an error. Now the object is
accessed directly. A side effect of this change is that if the object
has a library override and the active index is change, the modifier list
is displayed in the override color (turquoise, by default at least).